+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 - added subtotal to meter/offlinemeter/virtualmeter batch report 
 - added mqtt to data source protocol in myems-api
+- added tbl_template_files to myems_reporting_db
 ### Changed
 -